+++ app/AComp/Microservices.jsx
@@ -1,11 +1,31 @@
-import React, { useContext } from 'react';
+import React, { useContext, useState, useEffect } from 'react';
+import OverviewContext from '../context/OverviewContext';
 import HealthInformationContext from '../context/DetailsContext';
 import ServiceDetails from '../components/ServiceDetails.jsx';
+
 
 const { ipcRenderer } = window.require('electron');
 
 const Microservices = (props) => {
-  const { overviewState, setDetails, index } = props;
+  const { index, setDetails } = props;
+  // Overview state used to create service buttons
+  const [overviewState, setOverviewState] = useState([]);
+  // const [detailsSelected, setDetails] = useState();
+  // Contexts have data added to them following successful IPC return. Data is later used to create charts.
+  const serviceComponents = useContext(OverviewContext);
+
+  useEffect(() => {
+    // IPC communication used to initiate query for information on microservices.
+    ipcRenderer.send('overviewRequest', index);
+    // IPC listener responsible for retrieving infomation from asynchronous main process message.
+    ipcRenderer.on('overviewResponse', (event, data) => {
+      // Adds to state and context.
+      // console.log(JSON.parse(data));
+      setOverviewState(Object.values(JSON.parse(data)));
+      serviceComponents.overviewData = JSON.parse(data);
+    });
+  }, []);
+
   // Holds the buttons generated for unique services.
   const componentButtons = [];
 
@@ -26,6 +46,7 @@ const Microservices = (props) => {
             type="button"
             key={`serviceItem${index}${i}`}
             onClick={() => {
+              console.log(element.currentmicroservice);
               // IPC communication used to initiate query for information on microservice health information.
               ipcRenderer.send('detailsRequest', index);
 
@@ -35,12 +56,10 @@ const Microservices = (props) => {
                 healthdata.detailData = Object.values(JSON.parse(data));
                 // Updates state. Triggers rerender.
                 setDetails(
-                  <ServiceDetails
-                    service={element.currentmicroservice}
-                    setDetails={setDetails}
-                  />
+                      <ServiceDetails service={element.currentmicroservice} />
                 );
               });
+              // setSelection(<Monitoring detailsSelected={detailsSelected} />);
             }}
           >
             {element.currentmicroservice}
@@ -59,18 +78,17 @@ const Microservices = (props) => {
               type="button"
               key={`serviceItem${index}${i}`}
               onClick={() => {
-                ipcRenderer.send('detailsRequest', props.index);
-
+                ipcRenderer.send('detailsRequest', index);
+                
                 // IPC listener responsible for retrieving infomation from asynchronous main process message.
                 ipcRenderer.on('detailsResponse', (event, data) => {
                   // Adds returned data to context.
+                  console.log(element.currentmicroservice);
                   healthdata.detailData = Object.values(JSON.parse(data));
                   // Updates state. Triggers rerender.
                   setDetails(
-                    <ServiceDetails
-                      service={element.currentMicroservice}
-                      setDetails={setDetails}
-                    />
+                        <ServiceDetails service={element.currentmicroservice} />
+                      
                   );
                 });
               }}
