+++ app/components/ServiceDetails.jsx
@@ -5,7 +5,8 @@ import Modal from './Modal.jsx';
 // Renders charts created with health and communication data for a selected database.
 const ServiceDetails = (props) => {
   // Renders health info detail buttons
-  const { service, setDetails } = props;
+  const { service } = props;
+  console.log(service);
   // Hook used to toggle whether or not the Modal component renders
   const [modalDisplay, toggleModalDisplay] = useState(false);
   // Hook used to set the chart that the Modal displays.  The
@@ -85,7 +86,7 @@ const ServiceDetails = (props) => {
         onClick={() => {
           // document.location.reload()
           console.log('should clear data');
-          setDetails(null);
+          // setDetails(null);
         }}
       >
         Clear Health Data
