+++ app/containers/SideBarContainer.jsx
@@ -4,11 +4,10 @@ import Extras from '../AComp/Extras.jsx';
 import SetupContext from '../context/SetupContext';
 import DashboardContext from '../context/DashboardContext';
 import Microservices from '../AComp/Microservices.jsx';
-import OverviewContext from '../context/OverviewContext';
 import ServicesList from '../AComp/ServicesList.jsx';
 import AddService from '../components/AddService.jsx';
 import DeleteService from '../components/DeleteService.jsx';
-import Monitoring from './MonitoringContainer.jsx';
+// import Monitoring from './MonitoringContainer.jsx';
 import '../stylesheets/sidebar.css';
 
 const { ipcRenderer } = window.require('electron');
@@ -18,17 +17,15 @@ const SidebarContainer = (props) => {
   // Used to toggle setup required if user wants to add a new database.
   const setup = useContext(SetupContext);
 
-  const { setSelection } = props;
+  const { setSelection, setDetails} = props;
 
-  const [detailsSelected, setDetails] = useState();
+  
   // List of the databases saved by users to track microservices.
   const serviceList = useContext(DashboardContext);
 
-  // Overview state used to create service buttons
-  const [overviewState, setOverviewState] = useState([]);
+  
 
-  // Contexts have data added to them following successful IPC return. Data is later used to create charts.
-  const serviceComponents = useContext(OverviewContext);
+ 
   const [index, setIndex] = useState();
   const [isclicked, setClicked] = useState('false');
 
@@ -41,16 +38,8 @@ const SidebarContainer = (props) => {
   const ServicesClick = (e) => {
     clickToggle(e);
     setIndex(e.target.id);
-    // const event = e.target.id;
-    ipcRenderer.send('overviewRequest', e.target.id);
-    // IPC listener responsible for retrieving infomation from asynchronous main process message.
-    ipcRenderer.on('overviewResponse', (event, data) => {
-      // Adds to state and context.
-      // console.log(JSON.parse(data));
-      setOverviewState(Object.values(JSON.parse(data)));
-      serviceComponents.overviewData = JSON.parse(data);
-    });
-    setSelection(<Monitoring detailsSelected={detailsSelected} />);
+
+    // setSelection(<Monitoring detailsSelected={detailsSelected} />);
   };
   // Click function for AddService
   const AddClick = () => {
@@ -75,9 +64,9 @@ const SidebarContainer = (props) => {
         />
         {isclicked === 'true' ? (
           <Microservices
-            overviewState={overviewState}
-            setDetails={setDetails}
+            setSelection={setSelection}
             index={index}
+            setDetails={setDetails}
           />
         ) : null}
         <Extras
